refactor: remove BlockFetcher abstraction, restore viem client on provider
HyperSync only needs to replace preloading. Remove BlockFetcher
interface and RpcBlockFetcher — restore the original viem PublicClient
on EvmProvider for all live block operations. Only the Preloader
interface remains for optional HyperSync bulk fetching.
