+++ src/providers/evm/fetchers/index.ts
@@ -1,27 +1,22 @@
 import { CheckpointConfig } from '../../../types';
 import { Logger } from '../../../utils/logger';
-import { BlockFetcher, FetchedBlock, Preloader } from './types';
-import { RpcBlockFetcher } from './rpc';
+import { FetchedBlock, Preloader } from './types';
 import { HypersyncPreloader } from './hypersync';
 
-export { BlockFetcher, FetchedBlock, Preloader, RpcBlockFetcher, HypersyncPreloader };
+export { FetchedBlock, Preloader, HypersyncPreloader };
 
-export function createFetchers(
+export function createPreloader(
   config: CheckpointConfig,
   log: Logger
-): { fetcher: BlockFetcher; preloader?: Preloader } {
-  const fetcher = new RpcBlockFetcher(config.network_node_url);
-
+): Preloader | undefined {
   if (config.hypersync_api_token) {
     log.info('using HyperSync preloader');
 
-    const preloader = new HypersyncPreloader({
+    return new HypersyncPreloader({
       apiToken: config.hypersync_api_token,
       rpcUrl: config.network_node_url
     });
-
-    return { fetcher, preloader };
   }
 
-  return { fetcher };
+  return undefined;
 }
