+++ src/providers/evm/fetchers/types.ts
@@ -9,26 +9,6 @@ export type FetchedBlock = {
   timestamp: number;
 };
 
-export type BlockFetcher = {
-  getChainId(): Promise<number>;
-  getLatestBlockNumber(): Promise<number>;
-  getBlock(blockNumber: number): Promise<FetchedBlock>;
-  getBlockHash(blockNumber: number): Promise<string>;
-  getLogs(
-    fromBlock: number,
-    toBlock: number,
-    address: string | string[],
-    topics?: (string | string[])[]
-  ): Promise<Log[]>;
-  getLogsByBlockHash(blockHash: string): Promise<Log[]>;
-  getCheckpointsRange(
-    fromBlock: number,
-    toBlock: number,
-    sources: ContractSourceConfig[],
-    getEventHash: (name: string) => string
-  ): Promise<{ checkpoints: CheckpointRecord[]; logs: Log[] }>;
-};
-
 export type Preloader = {
   getCheckpointsRange(
     fromBlock: number,
