+++ src/providers/evm/provider.ts
@@ -1,19 +1,43 @@
 import {
+  createPublicClient,
+  formatLog,
   getAddress,
+  http,
   keccak256,
   Log,
   parseEventLogs,
   ParseEventLogsReturnType,
+  PublicClient,
+  RpcLog,
   stringToBytes
 } from 'viem';
-import { BlockFetcher, FetchedBlock, Preloader } from './fetchers/types';
+import { getRangeHint } from './helpers';
+import { FetchedBlock, Preloader } from './fetchers/types';
 import { Block, CustomJsonRpcError, EventsData, Writer } from './types';
 import { CheckpointRecord } from '../../stores/checkpoints';
 import { ContractSourceConfig } from '../../types';
+import { sleep } from '../../utils/helpers';
 import { BaseProvider, BlockNotFoundError, ReorgDetectedError } from '../base';
 
+type GetLogsBlockHashFilter = {
+  blockHash: string;
+};
+
+type GetLogsBlockRangeFilter = {
+  fromBlock: number;
+  toBlock: number;
+};
+
+/**
+ * Timeout for client requests in milliseconds.
+ * This timeout is also used when fetching latest blocks in getLogs.
+ */
+const CLIENT_TIMEOUT = 5 * 1000;
+
+const MAX_BLOCKS_PER_REQUEST = 10000;
+
 export class EvmProvider extends BaseProvider {
-  private readonly fetcher: BlockFetcher;
+  private readonly client: PublicClient;
   private readonly preloader?: Preloader;
 
   private readonly writers: Record<string, Writer>;
@@ -26,16 +50,19 @@ export class EvmProvider extends BaseProvider {
     log,
     abis,
     writers,
-    fetcher,
     preloader
   }: ConstructorParameters<typeof BaseProvider>[0] & {
     writers: Record<string, Writer>;
-    fetcher: BlockFetcher;
     preloader?: Preloader;
   }) {
     super({ instance, log, abis });
 
-    this.fetcher = fetcher;
+    this.client = createPublicClient({
+      transport: http(instance.config.network_node_url, {
+        timeout: CLIENT_TIMEOUT
+      })
+    });
+
     this.preloader = preloader;
     this.writers = writers;
   }
@@ -45,17 +72,23 @@ export class EvmProvider extends BaseProvider {
   }
 
   async getNetworkIdentifier(): Promise<string> {
-    const chainId = await this.fetcher.getChainId();
+    const chainId = await this.client.getChainId();
 
     return `evm_${chainId}`;
   }
 
   async getLatestBlockNumber(): Promise<number> {
-    return this.fetcher.getLatestBlockNumber();
+    const blockNumber = await this.client.getBlockNumber();
+
+    return Number(blockNumber);
   }
 
   async getBlockHash(blockNumber: number) {
-    return this.fetcher.getBlockHash(blockNumber);
+    const block = await this.client.getBlock({
+      blockNumber: BigInt(blockNumber)
+    });
+
+    return block.hash;
   }
 
   async processBlock(blockNumber: number, parentHash: string | null) {
@@ -78,13 +111,9 @@ export class EvmProvider extends BaseProvider {
             timestamp: BigInt(cached.timestamp)
           } as Block;
         } else {
-          const fetched = await this.fetcher.getBlock(blockNumber);
-          block = {
-            number: BigInt(fetched.number),
-            hash: fetched.hash,
-            parentHash: fetched.parentHash,
-            timestamp: BigInt(fetched.timestamp)
-          } as Block;
+          block = await this.client.getBlock({
+            blockNumber: BigInt(blockNumber)
+          });
         }
       }
     } catch (err) {
@@ -318,7 +347,9 @@ export class EvmProvider extends BaseProvider {
         throw new Error('Block hash is required to fetch logs from network');
       }
 
-      events = await this.fetcher.getLogsByBlockHash(blockHash);
+      events = await this._getLogs({
+        blockHash
+      });
     }
 
     return {
@@ -333,6 +364,134 @@ export class EvmProvider extends BaseProvider {
         return acc;
       }, {})
     };
+  }
+
+  /**
+   * This method is simpler implementation of getLogs method.
+   * This allows using two filters that are not supported in ethers v5:
+   * - `blockHash` to get logs for a specific block - if node doesn't know about that block it will fail.
+   * - `address` as a single address or an array of addresses.
+   * @param filter Logs filter
+   */
+  private async _getLogs(
+    filter: (GetLogsBlockHashFilter | GetLogsBlockRangeFilter) & {
+      address?: string | string[];
+      topics?: (string | string[])[];
+    }
+  ): Promise<Log[]> {
+    const params: {
+      fromBlock?: string;
+      toBlock?: string;
+      blockHash?: string;
+      address?: string | string[];
+      topics?: (string | string[])[];
+    } = {};
+
+    let signal: AbortSignal | undefined;
+
+    if ('blockHash' in filter) {
+      signal = AbortSignal.timeout(CLIENT_TIMEOUT);
+      params.blockHash = filter.blockHash;
+    }
+
+    if ('fromBlock' in filter) {
+      params.fromBlock = `0x${filter.fromBlock.toString(16)}`;
+    }
+
+    if ('toBlock' in filter) {
+      params.toBlock = `0x${filter.toBlock.toString(16)}`;
+    }
+
+    if ('address' in filter) {
+      params.address = filter.address;
+    }
+
+    if ('topics' in filter) {
+      params.topics = filter.topics;
+    }
+
+    const res = await fetch(this.instance.config.network_node_url, {
+      method: 'POST',
+      signal,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        jsonrpc: '2.0',
+        id: 1,
+        method: 'eth_getLogs',
+        params: [params]
+      })
+    });
+
+    if (!res.ok) {
+      throw new Error(`Request failed: ${res.statusText}`);
+    }
+
+    const json = await res.json();
+
+    if (json.error) {
+      throw new CustomJsonRpcError(
+        json.error.message,
+        json.error.code,
+        json.error.data
+      );
+    }
+
+    return json.result.map((log: RpcLog) => formatLog(log));
+  }
+
+  async getLogs(
+    fromBlock: number,
+    toBlock: number,
+    address: string | string[],
+    topics: (string | string[])[] = []
+  ): Promise<Log[]> {
+    let result = [] as Log[];
+
+    let currentFrom = fromBlock;
+    let currentTo = Math.min(toBlock, currentFrom + MAX_BLOCKS_PER_REQUEST);
+    while (true) {
+      try {
+        const logs = await this._getLogs({
+          fromBlock: currentFrom,
+          toBlock: currentTo,
+          address,
+          topics
+        });
+
+        result = result.concat(logs);
+
+        if (currentTo === toBlock) break;
+        currentFrom = currentTo + 1;
+        currentTo = Math.min(toBlock, currentFrom + MAX_BLOCKS_PER_REQUEST);
+      } catch (err: unknown) {
+        const rangeHint = getRangeHint(err, {
+          from: currentFrom,
+          to: currentTo
+        });
+
+        if (rangeHint) {
+          this.log.warn(
+            { err, rangeHint, fromBlock: currentFrom, toBlock: currentTo },
+            'getLogs failed. Received new range hint'
+          );
+
+          currentFrom = rangeHint.from;
+          currentTo = rangeHint.to;
+          continue;
+        }
+
+        this.log.error(
+          { fromBlock: currentFrom, toBlock: currentTo, address, err },
+          'getLogs failed'
+        );
+
+        await sleep(5000);
+      }
+    }
+
+    return result;
   }
 
   async getLogsForSources({
@@ -356,12 +515,9 @@ export class EvmProvider extends BaseProvider {
         source.events.map(event => this.getEventHash(event.name))
       );
 
-      const chunkEvents = await this.fetcher.getLogs(
-        fromBlock,
-        toBlock,
-        address,
-        [topics]
-      );
+      const chunkEvents = await this.getLogs(fromBlock, toBlock, address, [
+        topics
+      ]);
       events = events.concat(chunkEvents);
     }
 
@@ -373,7 +529,6 @@ export class EvmProvider extends BaseProvider {
     toBlock: number
   ): Promise<CheckpointRecord[]> {
     const sources = this.instance.getCurrentSources(toBlock);
-    const getEventHash = (name: string) => this.getEventHash(name);
 
     let checkpoints: CheckpointRecord[];
     let logs: Log[];
@@ -383,7 +538,7 @@ export class EvmProvider extends BaseProvider {
         fromBlock,
         toBlock,
         sources,
-        getEventHash
+        name => this.getEventHash(name)
       );
       checkpoints = result.checkpoints;
       logs = result.logs;
@@ -392,14 +547,17 @@ export class EvmProvider extends BaseProvider {
         this.blockCache.set(block.number, block);
       }
     } else {
-      const result = await this.fetcher.getCheckpointsRange(
+      const events = await this.getLogsForSources({
         fromBlock,
         toBlock,
-        sources,
-        getEventHash
-      );
-      checkpoints = result.checkpoints;
-      logs = result.logs;
+        sources
+      });
+
+      checkpoints = events.map(log => ({
+        blockNumber: Number(log.blockNumber),
+        contractAddress: log.address
+      }));
+      logs = events;
     }
 
     for (const log of logs) {
