+++ src/providers/evm/types.ts
@@ -7,7 +7,7 @@ import {
 } from 'viem';
 import { BaseWriterParams } from '../../types';
 
-export { BlockFetcher, FetchedBlock, Preloader } from './fetchers/types';
+export { FetchedBlock, Preloader } from './fetchers/types';
 
 export class CustomJsonRpcError extends Error {
   constructor(
