--- test/unit/providers/evm/fetchers/rpc.test.ts
@@ -1,8 +1,0 @@
-import { RpcBlockFetcher } from '../../../../../src/providers/evm/fetchers/rpc';
-
-describe('RpcBlockFetcher', () => {
-  it('should be instantiated with a network URL', () => {
-    const fetcher = new RpcBlockFetcher('https://rpc.example.com');
-    expect(fetcher).toBeDefined();
-  });
-});
